requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
kuzu==0.11.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
multiprocess==0.70.18
nltk==3.9.2
numpy==2.3.4
ollama==0.6.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
portalocker==3.2.0
protobuf==6.33.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
selenium==4.38.0
semchunk==3.2.5
sentry-sdk==2.43.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
starlette==0.49.3
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.2.0
